Move openapi.yaml to Schema/EIDO
This is a workaround for an issue encountered in the swagger/java parser.

If the parser is set to "resolve", it will resolve the NIEM person/vehicle schemas (and their dependencies). At this point, the NIEM types appear to be part of the EIDO schema. Unfortunately, since openapi.yaml in in the "Schema" directory and the NIEM components are in "Schema/NIEMComponents", the JSON-LD contexts (ex "../JSON-LD_Contexts/ncPersonNameTypeContext.jsonld") no longer work, since the context file is now relative to "./Schema" instead of "./Schema/NIEMContext".

To fix this, moved openapi.yaml file into a new Schema/EIDO directory.  Now relative paths work correctly for all components.